+++ packages/extention-contracts/src/interfaces/IMainMynaRegistrationVerifier.sol
@@ -3,9 +3,9 @@ pragma solidity ^0.8.19;
 
 interface IMainMynaRegistrationVerifier {
     function verifyProof(
-        uint[2] calldata _pA, 
-        uint[2][2] calldata _pB, 
-        uint[2] calldata _pC, 
-        uint[19] calldata _pubSignals
+        uint256[2] calldata _pA,
+        uint256[2][2] calldata _pB,
+        uint256[2] calldata _pC,
+        uint256[19] calldata _pubSignals
     ) external view returns (bool);
 }
