+++ packages/extention-contracts/src/interfaces/IMainMynaTree.sol
@@ -3,10 +3,10 @@ pragma solidity ^0.8.19;
 
 interface IMainMynaTree {
     function addMainAccount(
-        uint hashedModulus, 
-        uint hashedUserSecret,
-        uint[17] calldata signature,
-        uint[8] calldata proof
+        uint256 hashedModulus,
+        uint256 hashedUserSecret,
+        uint256[17] calldata signature,
+        uint256[8] calldata proof
     ) external;
 
     function getMerkleTreeRoot() external view returns (uint256);
@@ -14,9 +14,6 @@ interface IMainMynaTree {
     function getMerkleTreeDepth() external view returns (uint256);
 
     function getNumberOfMerkleTreeLeaves() external view returns (uint256);
-    
-    function verifyMainAccount(
-        uint[8] calldata proof,
-        uint identityCommitment
-    ) external view returns (bool);
+
+    function verifyMainAccount(uint256[8] calldata proof, uint256 identityCommitment) external view returns (bool);
 }
